Refactor React state management and props handling
