+++ package.json
@@ -10,7 +10,8 @@
   "scripts": {
     "react-dev": "webpack -d -p --watch",
     "server-dev": "nodemon ./server/index.js",
-    "write-csv": "node --max-old-space-size=8192 ./server/db/writecsv.js",
+    "write-csv-psql": "node --max-old-space-size=8192 ./server/db/write-csv-psql.js",
+    "write-csv-cql": "node --max-old-space-size=8192 ./server/db/write-csv-cql.js",
     "seed-db": "psql -d postgres -a -f pschema.sql",
     "test": "jest --Coverage --cache",
     "lint": "./node_modules/.bin/eslint ./client"
@@ -19,6 +20,7 @@
     "axios": "^0.19.0",
     "bluebird": "^3.5.5",
     "body-parser": "^1.19.0",
+    "cassandra-driver": "^4.1.0",
     "cors": "^2.8.5",
     "csv-writer": "^1.5.0",
     "express": "^4.17.1",
